provide random number generators for seeding db
functions for prices and stock quantities
decided to set default values for unit, seedsPerUnit attributes to avoid
having to randomize that.
set stock to allowNull: false because we really should always have a
stock quantity on file.
set up pending specs for the random number generation portion of
seeding just in case we really want to automate testing that portion.