refactor: break errorHandler into its own middleware
If you go to the `/api/asdfasdfa` nonsense path, this will return with a
404 status code and the text 'Oops... resource not found.' This does not
resolve the original change request for a true 404 page that works for
all possible request URIs. However, it does make it easier for us to
possibly incorporate a real solution in the future such as server-side
rendering.